+++ protected/humhub/modules/user/tests/codeception/unit/MentionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace tests\codeception\unit\modules\space;
+namespace humhub\modules\user\tests\codeception\unit;
 
 use tests\codeception\_support\HumHubDbTestCase;
 use Codeception\Specify;
@@ -30,10 +30,11 @@ class MentionTest extends HumHubDbTestCase
         $this->assertMailSent(1, 'Mentioned Notification');
     }
     
-    public function testCreateComment()
+    public function testMentionAuthor()
     {
         $this->becomeUser('User2');
         
+        // Mention Admin in Space 1 (Admin is author of post)
         $comment = new Comment([
             'message' => 'Hi @-u01e50e0d-82cd-41fc-8b0c-552392f5839c',
             'object_model' => Post::className(),
@@ -43,7 +44,27 @@ class MentionTest extends HumHubDbTestCase
         $comment->save();
         
         $this->assertHasNotification(Mentioned::class, $comment);
-        // Commented and Mentioned mail
+        
+        // We expect only the Mentioned mail
+        $this->assertMailSent(1, 'Comment Notification Mail sent');
+    }
+    
+    public function testMentionNonAuthor()
+    {
+        $this->becomeUser('User2');
+        
+        // Mention User1 in post
+        $comment = new Comment([
+            'message' => 'Hi @-u01e50e0d-82cd-41fc-8b0c-552392f5839d',
+            'object_model' => Post::className(),
+            'object_id' => 7
+        ]);
+
+        $comment->save();
+        
+        $this->assertHasNotification(Mentioned::class, $comment);
+        
+        // Commented mail for Admin and Mentioned mail for User1
         $this->assertMailSent(2, 'Comment Notification Mail sent');
     }
 }